% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filters.R
\name{applyFilter}
\alias{applyFilter}
\title{Filtering data}
\usage{
applyFilter(x, type)
}
\arguments{
\item{x}{An eapr object}

\item{type}{Filter or list of filters to apply to dataset. Currently, only the filter used in Fama French 1992 is implemented.}
}
\value{
A filtered EAPR object
}
\description{
Filtering data
}
