% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{extract}
\alias{extract}
\title{Access the wrds database}
\usage{
extract(username, src = "wrds", fundamental.var = c("ME", "BE",
  "BE/ME", "A/ME", "A/BE", "OP", "INV", "E/P", "CF/P", "D/P"),
  from = as.Date("1963-07-31"), to = as.Date(paste0(format(Sys.Date() -
  years(1), "\%Y"), "-12-31")), periodicity = "M",
  rebalance.freq = "A", drop.excess = T, preceding = 60,
  min.prec = 0.4)
}
\arguments{
\item{username}{username for database}

\item{src}{source of database}

\item{fundamental.var}{fundamental variables to be computed/extracted}

\item{from}{starting date}

\item{to}{end date}

\item{periodicity}{frequency of the data}

\item{rebalance.freq}{how often portfolios are rebalanced}

\item{drop.excess}{Drop variables used to calculate fundamentals and technicals}

\item{preceding}{number of preceding periods to consider for technical variables}

\item{min.prec}{minimum number of preceding periods necessary to compute technical variables}
}
\value{
An eapr object containing the merged crsp and compustat data (ccm), the time series of returns for the index and each security (market.dt),
}
\description{
This function accesses the wrds database and returns
}
