% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eapr_lm.R
\name{FamaMacbeth}
\alias{FamaMacbeth}
\alias{summary.FamaMacbeth}
\alias{plot.FamaMacbeth}
\title{Fama MacBeth Cross-Sectional Regressions}
\usage{
FamaMacbeth(formula, x, type = "ols", outliers.method = "winsorize",
  outliers.level = 0.005, robust.control = NULL)

\method{summary}{FamaMacbeth}(x)

\method{plot}{FamaMacbeth}(x)
}
\arguments{
\item{formula}{An eapr object which contains the time series of variables for all
stocks.}

\item{x}{Either a vector of percentile cut points or the number of quantiles
to compute. If the latter, quantiles are computed using evenly spaced percentile
cut points.}

\item{type}{Type(s) of regression method(s) to use. Current supported methods are least-squares and robust MM regressions. 
Least-squares may be selected by giving this parameter a value 'ols' or 'ls' and Robust MM is chosen by giving this parameter a value of 'mm' or 
'robust.mm'.}

\item{outliers.method}{Method for dealing with outliers. Options are 'trim', 'winzorize', and 'none'.}

\item{outliers.level}{Level to use for outliers. Indicates that whichever method chosen, unless 'none', will be applied to the x and 1 - x percentiles.}

\item{robust.control}{Specifies model parameters for robust regression. See \code{\link[RobStatTM]{lmrobdet.control}} for more info.}
}
\value{
Data.table containing a time series of intercepts and slopes if only one regression method is selected. A list of data.tables if multiple
regression methods selected.
}
\description{
Computes cross-sectional regressions for each period
}
\author{
Gregory Brownson, \email{gregory.brownson@gmail.com}
}
