% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filters.R
\name{ff92Filter}
\alias{ff92Filter}
\title{Filtering data using methods detailed in Fama French (1992)}
\usage{
ff92Filter(x)
}
\arguments{
\item{x}{An eapr object}
}
\value{
A filtered EAPR object using the same filtering methods used in Fama French (1992)
}
\description{
Filtering data using methods detailed in Fama French (1992)
}
