% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eapr_portfolios.R
\name{quantilePortfolio.eapr}
\alias{quantilePortfolio.eapr}
\title{Quantile portfolio formation}
\usage{
\method{quantilePortfolio}{eapr}(x, q, on, sort = "uni")
}
\arguments{
\item{x}{an eapr object which contains the time series of variables for all
stocks.}

\item{q}{Either a vector of percentile cut points or the number of quantiles
to compute. If the latter, quantiles are computed using evenly spaced percentile
cut points.}
}
\description{
This function creates a univariate or bivariate quantile portfolio
}
