% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eapr_portfolios.R
\name{quantilePortfolio}
\alias{quantilePortfolio}
\alias{quartilePortfolio}
\alias{quintilePortfolio}
\alias{decilePortfolio}
\title{Quantile portfolio formation}
\usage{
quantilePortfolio(x, q, on.x, on.y = NULL, sort = "univariate")

quartilePortfolio(x, on.x, on.y = NULL, sort = "univariate")

quintilePortfolio(x, on.x, on.y = NULL, sort = "univariate")

decilePortfolio(x, on.x, on.y = NULL, sort = "univariate")
}
\arguments{
\item{x}{An eapr object which contains the time series of variables for all
stocks.}

\item{q}{Either a vector of percentile cut points or the number of quantiles
to compute. If the latter, quantiles are computed using evenly spaced percentile
cut points.}

\item{on.x}{First variable to sort on.}

\item{on.y}{Second variable to sort on. This option is ignored if 'univariate' sorting is selected.}

\item{sort}{Type of sort for portfolios. Options are univariate, bivariate.ind
(independent bivariate), and bivariate.dep (dependent bivariate) sorts.}
}
\value{
An eaprPortfolio object containing a data.table of the time series of portfolio assignments for each stock in each period.
}
\description{
This function creates a univariate or bivariate quantile portfolio
}
