% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filters.R
\name{applyFilter}
\alias{applyFilter}
\title{applyFilter}
\usage{
applyFilter(x, type)
}
\arguments{
\item{x}{An eapr object}

\item{type}{Filter or list of filters to apply to data}
}
\description{
Implementation of filter outlined in Fama and French (1992).
}
