% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filters.R
\name{applyFilter}
\alias{applyFilter}
\title{filter}
\usage{
applyFilter(x, type)
}
\arguments{
\item{x}{eapr object}

\item{type}{Single filter or list of filters}
}
\description{
Implementation of filter outlined in Fama and French (1992).
}
