% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{extract}
\alias{extract}
\title{Access the wrds database}
\usage{
extract(username, src = "wrds", variables = c("ME", "BE", "BE/ME",
  "A/ME", "A/BE", "OP", "INV", "E/P", "CF/P", "D/P"),
  from = as.Date("1963-07-31"), to = as.Date(paste0(format(Sys.Date() -
  years(1), "\%Y"), "-12-31")), periodicity = "M",
  rebalance.freq = "A", drop.excess = T, preceding = 60,
  min.prec = 0.4)
}
\arguments{
\item{min.prec}{}
}
\description{
This function accesses the wrds database and returns
}
